move toroidal coord transforms to tools.py
